+++ src/extensions/mainloop/mainloopext_hooks.cpp
@@ -29,12 +29,13 @@
 #include "vinifera_globals.h"
 #include "tibsun_globals.h"
 #include "tibsun_functions.h"
+#include "command.h"
 #include "iomap.h"
 #include "tactical.h"
 #include "house.h"
 #include "fatal.h"
 #include "debughandler.h"
-#include "asserthandler.h"
+#include "asserthandler.h"6
 
 #include "hooker.h"
 #include "hooker_macros.h"
@@ -175,6 +176,56 @@ static bool Main_Loop_Intercept()
     return ret;
 }
 
+
+void Process_Command_If_Allowed(CommandClass* command)
+{
+    if (!Scen->UserInputLocked || (CommandClass::From_Type(COMMAND_OPTIONS) == command)) {
+        command->Process();
+    }
+}
+
+/**
+ *  #issue-255
+ *
+ *  Fixes the user being able to do keyboard input and as such, affect
+ *  game state while input is locked.
+ *
+ *  @author: Rampastring
+ */
+DECLARE_PATCH(_Main_Loop_Check_Keyboard_Input_Allowed)
+{
+    GET_REGISTER_STATIC(CommandClass*, command, ecx);
+    Process_Command_If_Allowed(command);
+    JMP(0x00508EA8);
+}
+
+DECLARE_PATCH(_Keyboard_Process_Check_Keyboard_Input_Allowed)
+{
+    GET_REGISTER_STATIC(CommandClass*, command, ecx);
+    Process_Command_If_Allowed(command);
+
+    // Rebuild function epilogue, we destroyed one byte of it
+    // by jumping to this hack
+    _asm { pop esi }
+    _asm { pop ebp }
+    _asm { retn }
+}
+
+DECLARE_PATCH(_Sync_Delay_Check_Keyboard_Input_Allowed_Patch1)
+{
+    GET_REGISTER_STATIC(CommandClass*, command, eax);
+    Process_Command_If_Allowed(command);
+    JMP(0x00509659);
+}
+
+DECLARE_PATCH(_Sync_Delay_Check_Keyboard_Input_Allowed_Patch2)
+{
+    GET_REGISTER_STATIC(CommandClass*, command, ecx);
+    Process_Command_If_Allowed(command);
+    JMP(0x0050976C);
+}
+
+
 /**
  *  Main function for patching the hooks.
  */
@@ -184,4 +235,16 @@ void MainLoop_Hooks()
     Patch_Call(0x00462A9C, &Main_Loop_Intercept);
     Patch_Call(0x005A0B85, &Main_Loop_Intercept);
     Patch_Jump(0x005B10F0, &_Queue_Options_Frame_Step_Check_Patch);
+
+    /**
+     *  #issue-255
+     *
+     *  Keyboard processing patches.
+     *  Technically not all are directly in the main loop,
+     *  but all of these are similar to the main loop patch so we include them all here.
+     */
+    Patch_Jump(0x00508E83, &_Main_Loop_Check_Keyboard_Input_Allowed);
+    Patch_Jump(0x0050945C, &_Keyboard_Process_Check_Keyboard_Input_Allowed);
+    Patch_Jump(0x00509632, &_Sync_Delay_Check_Keyboard_Input_Allowed_Patch1);
+    Patch_Jump(0x00509747, &_Sync_Delay_Check_Keyboard_Input_Allowed_Patch2);
 }